requirements.txt
# Core FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
email-validator==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Redis & Celery (Phase 4)
redis==5.0.1
celery==5.3.6
flower==2.0.1

# HTTP clients
httpx==0.26.0
aiohttp==3.11.10           # Updated to latest

# AI & NLP (Phase 2)
anthropic==0.18.1
tiktoken==0.5.2
textstat==0.7.3

# Web scraping & Adapters (Phase 3)
beautifulsoup4==4.12.3
lxml==5.3.0                # Updated to latest
feedparser==6.0.11
playwright==1.48.0         # Updated to latest
markdown==3.7              # Markdown to HTML conversion

# Utilities
python-dateutil==2.8.2
pytz==2024.1
croniter==2.0.1
python-slugify==8.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
